Enhance error handling and auth
- Add error handler middleware with request ID logging
- Refactor auth forms: remove Card wrapper, centralize error handling
- Move SignIn page to separate component
- Add minimal development script
- Update error display utilities and imports
