+++ apps/dockstat/src/components/auth/LocalLoginForm.tsx
@@ -1,18 +1,19 @@
-import { Card, Input } from "@dockstat/ui"
+import { Input } from "@dockstat/ui"
 import { ArrowRight, Eye, EyeOff, Loader2, Shield } from "lucide-react"
 import { useLocalLogin } from "@/hooks/useLocalLogin"
 
-export function LocalLoginForm({ show }: { show: boolean }) {
-  if (!show) return
-
-  const { formData, error, isSubmitting, showPassword, handleSubmit, updateField, togglePassword } =
-    useLocalLogin()
+export function LocalLoginForm({
+  error,
+  setError,
+}: {
+  error: string | null
+  setError: (err: string | null) => void
+}) {
+  const { formData, isSubmitting, showPassword, handleSubmit, updateField, togglePassword } =
+    useLocalLogin({ error, setError })
 
   return (
-    <Card
-      className="p-5 sm:p-6 rounded-2xl"
-      glass
-    >
+    <>
       <div className="flex items-center gap-3 mb-5">
         <div className="w-9 h-9 rounded-lg flex items-center justify-center bg-indigo-500/10 border border-indigo-500/20">
           <Shield
@@ -36,7 +37,7 @@ export function LocalLoginForm({ show }: { show: boolean }) {
           </p>
           <div className="field-shell px-4 py-3">
             <Input
-              className="!bg-transparent !border-0 !shadow-none !p-0 !ring-0 w-full text-sm text-white/80 placeholder:text-white/25"
+              className="bg-transparent! border-0! shadow-none! p-0! ring-0! w-full text-sm text-white/80 placeholder:text-white/25"
               disabled={isSubmitting}
               onChange={(v) => updateField("name", v)}
               placeholder="Enter your username"
@@ -51,7 +52,7 @@ export function LocalLoginForm({ show }: { show: boolean }) {
           </p>
           <div className="field-shell px-4 py-3 flex items-center">
             <Input
-              className="!bg-transparent !border-0 !shadow-none !p-0 !ring-0 w-full text-sm text-white/80 placeholder:text-white/25"
+              className="bg-transparent! border-0! shadow-none! p-0! ring-0! w-full text-sm text-white/80 placeholder:text-white/25"
               disabled={isSubmitting}
               onChange={(v) => updateField("pass", v)}
               placeholder="Enter your password"
@@ -59,7 +60,7 @@ export function LocalLoginForm({ show }: { show: boolean }) {
               value={formData.pass}
             />
             <button
-              className="flex-shrink-0 ml-2 p-1 rounded-md text-white/20 hover:text-white/40 transition-colors"
+              className="shrink-0 ml-2 p-1 rounded-md text-white/20 hover:text-white/40 transition-colors"
               onClick={togglePassword}
               tabIndex={-1}
               type="button"
@@ -68,12 +69,6 @@ export function LocalLoginForm({ show }: { show: boolean }) {
             </button>
           </div>
         </div>
-
-        {error && (
-          <div className="px-4 py-3 rounded-xl bg-red-500/5 border border-red-500/10">
-            <p className="text-sm text-red-300/80">{error}</p>
-          </div>
-        )}
 
         <button
           className="cta-button w-full py-3.5 text-white font-semibold text-sm flex items-center justify-center gap-2"
@@ -96,6 +91,6 @@ export function LocalLoginForm({ show }: { show: boolean }) {
           )}
         </button>
       </form>
-    </Card>
+    </>
   )
 }
