+++ apps/dockstat/src/components/auth/constants.tsx
@@ -1,0 +1,80 @@
+import {
+  SiAuth0,
+  SiAuthelia,
+  SiAuthentik,
+  SiBitbucket,
+  SiDiscord,
+  SiFacebook,
+  SiFusionauth,
+  SiGithub,
+  SiGitlab,
+  SiGoogle,
+  SiKeycloak,
+  SiOkta,
+  SiSpotify,
+  SiTwitch,
+  SiYandexcloud,
+} from "@icons-pack/react-simple-icons"
+
+export const floatingIcons = [
+  <SiAuthelia
+    className="w-full h-full"
+    key="authelia"
+  />,
+  <SiKeycloak
+    className="w-full h-full"
+    key="keycloak"
+  />,
+  <SiAuthentik
+    className="w-full h-full"
+    key="authentik"
+  />,
+  <SiOkta
+    className="w-full h-full"
+    key="okta"
+  />,
+  <SiGoogle
+    className="w-full h-full"
+    key="google"
+  />,
+  <SiGithub
+    className="w-full h-full"
+    key="github"
+  />,
+  <SiBitbucket
+    className="w-full h-full"
+    key="bitbucket"
+  />,
+  <SiDiscord
+    className="w-full h-full"
+    key="discord"
+  />,
+  <SiFacebook
+    className="w-full h-full"
+    key="facebook"
+  />,
+  <SiGitlab
+    className="w-full h-full"
+    key="gitlab"
+  />,
+  <SiSpotify
+    className="w-full h-full"
+    key="spotify"
+  />,
+  <SiTwitch
+    className="w-full h-full"
+    key="twitch"
+  />,
+  <SiAuth0
+    className="w-full h-full"
+    key="auth0"
+  />,
+  <SiFusionauth
+    className="w-full h-full"
+    key="fusionauth"
+  />,
+  <SiYandexcloud
+    className="w-full h-full"
+    key="yandexcloud"
+  />,
+]
