+++ apps/dockstat/src/pages/auth/Callback.tsx
@@ -1,9 +1,9 @@
 import { Button, Card, CardBody, CardHeader } from "@dockstat/ui"
 import { useContext, useEffect, useState } from "react"
 import { useNavigate } from "react-router"
+import { floatingIcons } from "@/components/auth/constants"
 import { AnimatedIconBackground } from "@/components/auth/SignInBg"
 import { EdenClientContext } from "@/contexts/edenClient"
-import { floatingIcons } from "../SignIn"
 
 const API_BASE_URL = "http://localhost:3030/api/v2"
